requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Mail==0.9.1
openai>=1.12.0
python-dotenv==1.0.0
werkzeug==2.3.7
pytest==7.4.2
PyJWT==2.8.0
Flask-Talisman==1.1.0
Flask-Limiter==3.5.1
Flask-Migrate==4.0.5
itsdangerous==2.1.2
